Refactored buttons - needed for implementation of tournament sign up
